requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
python-dotenv==1.0.1
httpx==0.26.0
websockets==12.0
pandas==2.2.0
numpy==1.26.3
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-telegram-bot==20.7
cryptography==42.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
